Streamline structure of Rod 1D problem formulations.
