+++ engioptiqa/problems/rod_1d/__init__.py
@@ -1,9 +1,9 @@
-from .design_optimization_problem_amplify import DesignOptimizationProblemAmplify
+from .design_optimization_problem import DesignOptimizationProblem
 from .rod_1d import Rod1D
-from .structural_analysis_problem_amplify import StructuralAnalysisProblemAmplify
+from .structural_analysis_problem import StructuralAnalysisProblem
 
 __all__ = [
-    "DesignOptimizationProblemAmplify",
+    "DesignOptimizationProblem",
     "Rod1D",
-    "StructuralAnalysisProblemAmplify",
+    "StructuralAnalysisProblem",
 ]
