+++ engioptiqa/problems/rod_1d/design_optimization_problem.py
@@ -1,4 +1,5 @@
-from abc import abstractmethod
+# from abc import abstractmethod
+from amplify import QuadratizationMethod
 import itertools
 from matplotlib import pyplot as plt
 import matplotlib.patches as patches
@@ -58,10 +59,34 @@ class DesignOptimizationProblem(BaseProblem):
 
         output = f'Analytic Force: {self.force_analytic}\n'
         self.print_and_log(output)
-        
-    @abstractmethod
-    def generate_cross_section_polys(self):
-        pass
+
+    def generate_discretization(self, n_qubits_per_node, binary_representation):
+        BaseProblem.initialize_discretization(self)
+        BaseProblem.generate_nodal_force_polys(self, n_qubits_per_node, binary_representation)
+        self.generate_cross_section_inverse_polys()
+
+    def generate_cross_section_inverse_polys(self):
+        assert(self.symbol_generator is not None)
+        cs_inv_polys = []
+
+        for _ in range(self.rod.n_comp):
+            q = self.symbol_generator.array(1)
+            cs_inv_polys.append(1./self.A_choice[0] + (1./self.A_choice[1]-1./self.A_choice[0])*q[0])
+        self.cs_inv_polys = cs_inv_polys
+
+    def set_quad_method(self,quad_method_name):
+        self.quad_method_name = quad_method_name
+        match quad_method_name:
+            case 'ISHIKAWA':
+                self.quad_method = QuadratizationMethod.ISHIKAWA
+            case 'ISHIKAWA_KZFD':
+                self.quad_method = QuadratizationMethod.ISHIKAWA_KZFD
+            case 'SUBSTITUTION':
+                self.quad_method = QuadratizationMethod.SUBSTITUTION
+            case 'SUBSTITUTION_KZFD':
+                self.quad_method = QuadratizationMethod.SUBSTITUTION_KZFD
+            case _ :
+                raise Exception('Unknown quad_method', quad_method_name)
 
     def visualize_qubo_matrix_pattern(self, show_fig=False, save_fig=False, save_tikz=False, suffix=''):
         self.plot_qubo_matrix_pattern()
